[ feat ] : 폰트 스타일 테스트, 세로 높이 수정
