+++ components/common/Layout.tsx
@@ -27,12 +27,15 @@ export default Layout;
 const LayoutWrapper = styled.div`
   display: flex;
   justify-content: center;
+  height: 100vh;
 `;
 const ContentWrapper = styled.div`
   display: flex;
+
   flex-wrap: wrap;
 
   width: 37.5rem;
+  height: 100%;
   min-height: 81.2rem;
   background: ${theme.colors.backgroundColor};
 `;
