+++ pages/calendar/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import TupleCalendar from 'components/TupleCalendar';
+import TupleCalendar from 'components/calendar/TupleCalendar';
 
 const index = () => {
   return <TupleCalendar />;
