+++ public/fonts/style.css
@@ -1,0 +1,5 @@
+@font-face {
+    font-family: 'Noto Sans KR';
+    src: url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@400;500;700&display=swap');
+    font-style: normal;
+}
