+++ src/styles/globalStyle.ts
@@ -4,11 +4,13 @@ import reset from 'styled-reset';
 const GlobalStyle = createGlobalStyle`
   ${reset};
   
-  html,
-  body {
-    width: 100%;
-    height: 100%;
-  }
+    html,
+    body {
+    max-width: 100vw;
+    min-height:100vh;
+    overflow-x: hidden;
+    }
+
   
   #root {
     margin: 0 auto;
@@ -18,10 +20,17 @@ const GlobalStyle = createGlobalStyle`
     font-size: 62.5%;
   }
   
-  * {
+ * {
     box-sizing: border-box;
-  }
+    padding: 0;
+    margin: 0;
+    }
   
+
+  ul, ol {
+  list-style: none;
+  }
+
   body, button {
     font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans,
     Helvetica Neue, sans-serif;
